+++ MoMoney.Core/Helpers/PageLoader.cs
@@ -6,8 +6,6 @@ public static class PageLoader
     public static async Task Load(Func<Task> targetMethod)
     {
         await Task.Delay(400);
-        Shell.Current.IsBusy = true;
         await targetMethod();
-        Shell.Current.IsBusy = false;
     }
 }